+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Add `clusteradm delete token` [Create cmd to revoke the bootstrap token](https://github.com/open-cluster-management-io/clusteradm/issues/23)
 - Add the capability of injecting rendering functions [While rendering a yaml I want to inject my own functions](https://github.com/open-cluster-management-io/clusteradm/issues/37)
 - [Enable external cluster config to support oidc](https://github.com/open-cluster-management-io/clusteradm/issues/38)
+- Add klog flags [Log will be helpful to contributors](https://github.com/open-cluster-management-io/clusteradm/issues/36)
+
 ## Breaking Changes
 
 ## Changes